Remove unused imports from the linter Refactor into a frame
