+++ package.json
@@ -28,6 +28,7 @@
         "electron-is-dev": "^1.2.0",
         "eslint": "^7.23.0",
         "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
+        "eslint-plugin-unused-imports": "^1.1.1",
         "html-webpack-plugin": "^5.1.0",
         "jest": "^26.6.3",
         "nodemon": "^2.0.7",
